requirements_win.txt
# Windows-specific requirements for AI Governance Experiment
# Base requirements
transformers>=4.31.0
accelerate>=0.21.0
numpy>=1.24.3
tqdm>=4.65.0
pyyaml>=6.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# PyTorch - CPU version (add --index-url https://download.pytorch.org/whl/cpu when installing)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Oumi for consistent API
oumi>=0.1.7

# Alternative to llama.cpp for Windows
sentencepiece>=0.1.99
ctransformers>=0.2.24
optimum>=1.12.0

# NLP tools for analysis
nltk>=3.8.1
scikit-learn>=1.2.2
spacy>=3.6.0

# Plotting and visualization
plotly>=5.14.0